Refactor mainNew into modules with logging
