+++ bot/state.py
@@ -1,0 +1,4 @@
+# Global runtime state for the bot
+voice_times = {}
+user_dm_counts = {}
+invite_cache = {}
